[RHTAP-1321] Moving methods
