+++ pkg/framework/framework.go
@@ -136,10 +136,7 @@ func InitControllerHub(cc *kubeCl.CustomClient) (*ControllerHub, error) {
 		return nil, err
 	}
 	// Initialize Tekton controller
-	tektonController, err := tekton.NewSuiteController(cc)
-	if err != nil {
-		return nil, err
-	}
+	tektonController := tekton.NewSuiteController(cc)
 
 	// Initialize GitOps controller
 	gitopsController, err := gitops.NewSuiteController(cc)
